requirements.txt
pandas==2.1.1
httpx==0.26.0
python-dotenv==1.0.0
uvicorn==0.15.0
fastapi
geopy
